requirements.txt
absl-py==1.4.0
aniso8601==9.0.1
appnope==0.1.3
asttokens==2.2.1
astunparse==1.6.3
attrs==22.1.0
backcall==0.2.0
beautifulsoup4==4.11.1
bs4==0.0.1
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
configobj==5.0.6
contourpy==1.0.7
cryptography==39.0.1
cycler==0.11.0
decorator==5.1.1
executing==1.2.0
Flask==2.1.2
Flask-API==1.0
Flask-RESTful
Flask-Cors==3.0.10
Flask-JWT-Extended==4.4.1
flask-restplus==0.10.1
flask-restx==0.5.1
Flask-Script==2.0.6
Flask-SQLAlchemy==2.5.1
flatbuffers==23.1.21
fonttools==4.38.0
gast==0.4.0
google-auth==2.16.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
greenlet==2.0.1
grpcio==1.51.1
gunicorn==20.1.0
h5py==3.8.0
idna==3.4
importlib-metadata==5.1.0
importlib-resources==5.10.1
iniconfig==1.1.1
ipython==8.9.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.17.3
keras==2.11.0
kiwisolver==1.4.4
libclang==15.0.6.1
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.6.3
matplotlib-inline==0.1.6
numpy==1.24.1
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==22.0
pandas==1.5.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
pkgutil_resolve_name==1.3.10
pluggy==1.0.0
prompt-toolkit==3.0.36
protobuf==3.19.6
ptyprocess==0.7.0
pure-eval==0.2.2
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
Pygments==2.14.0
PyJWT==2.6.0
PyMySQL==1.0.2
pyparsing==3.0.9
pypng==0.20220715.0
pyrsistent==0.19.2
pytest==7.1.2
python-dateutil==2.8.2
python-decouple==3.7
python-dotenv==0.19.2
pytz==2022.6
qrcode==7.4.2
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.2.1
scipy==1.10.0
seaborn==0.12.2
six==1.16.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.46
stack-data==0.6.2
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow-estimator==2.11.0
termcolor==2.2.0
threadpoolctl==3.1.0
tomli==2.0.1
traitlets==5.9.0
typing_extensions==4.4.0
urllib3==1.26.13
wcwidth==0.2.6
Werkzeug==2.1.2
wrapt==1.14.1
zipp==3.11.0
